Removed templategallery webview
