+++ .vscodeignore
@@ -14,7 +14,6 @@ dist/test/**
 docs/**
 gulp*
 node_modules/**
-!node_modules/@vscode/codicons/**
 out/**
 resources/readme/**
 src/**
