+++ package.json
@@ -2,7 +2,7 @@
     "name": "vscode-azurefunctions",
     "displayName": "Azure Functions",
     "description": "%azureFunctions.description%",
-    "version": "1.20.4-alpha.90",
+    "version": "1.21.0",
     "publisher": "ms-azuretools",
     "icon": "resources/azure-functions.png",
     "aiKey": "0c6ae279ed8443289764825290e4f9e2-1a736e7c-1324-4338-be46-fc2a58ae4d14-7255",
@@ -492,6 +492,8 @@
                 "category": "Azure Functions",
                 "icon": "$(run-all)",
                 "enablement": "!virtualWorkspace"
+            },
+            {
                 "command": "azureFunctions.funcHostDebug.refresh",
                 "title": "%azureFunctions.funcHostDebug.refresh%",
                 "category": "Azure Functions",
@@ -1637,14 +1639,15 @@
     },
     "scripts": {
         "vscode:prepublish": "npm run build",
-        "build": "npm run build:esbuild & npm run build:check",
+        "build": "npm run build:esbuild & npm run build:webview & npm run build:check",
         "build:esbuild": "node esbuild.mjs",
+        "build:webview": "node scripts/copyWebviewAssets.mjs",
         "build:check": "tsc --noEmit",
         "check": "tsc --noEmit",
         "compile": "tsc -watch",
         "version": "tsc --version",
         "cleanReadme": "node scripts/cleanReadme.mjs",
-        "package": "vsce package --githubBranch main",
+        "package": "vsce package --githubBranch main --no-dependencies",
         "lint": "eslint --max-warnings 0",
         "lint-fix": "eslint --max-warnings 0 --fix",
         "pretest": "npm run build && node scripts/pretest.mjs",
@@ -1654,6 +1657,7 @@
     "devDependencies": {
         "@azure/arm-msi": "^2.1.0",
         "@azure/arm-resources": "^5.2.0",
+        "@microsoft/applicationinsights-common": "^3.4.1",
         "@microsoft/vscode-azext-eng": "^1.0.0-alpha.16",
         "@types/fs-extra": "^8.1.0",
         "@types/node": "^22.0.0",
@@ -1687,9 +1691,9 @@
         "@microsoft/vscode-azext-azureappsettings": "^1.0.0",
         "@microsoft/vscode-azext-azureutils": "^4.0.1",
         "@microsoft/vscode-azext-utils": "^4.0.5",
+        "@microsoft/vscode-azext-webview": "file:../vscode-azuretools/webview",
         "@microsoft/vscode-azureresources-api": "^2.6.2",
         "@microsoft/vscode-container-client": "^0.3.0",
-        "@vscode/codicons": "^0.0.44",
         "@microsoft/vscode-processutils": "^0.2.1",
         "cross-fetch": "^4.0.0",
         "escape-string-regexp": "^4.0.0",
